requirements-py312.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
bottle==0.13.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
distro==1.9.0
dmgbuild==1.6.7
docstring_parser==0.18.0
ds_store==1.3.2
fastapi==0.136.1
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
jieba==0.42.1
jiter==0.14.0
mac_alias==2.2.3
macholib==1.16.4
modulegraph==0.19.7
openai==2.33.0
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
proxy_tools==0.1.0
py2app==0.28.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-UserNotifications==12.1
pyobjc-framework-WebKit==12.1
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
pywebview==6.2.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
setuptools==82.0.1
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.47.0
